Move ./codes -> ./trace/codes
