+++ sdk/trace/trace_test.go
@@ -26,10 +26,10 @@ import (
 	"time"
 
 	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/label"
 	"go.opentelemetry.io/otel/oteltest"
 	"go.opentelemetry.io/otel/trace"
+	"go.opentelemetry.io/otel/trace/codes"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/assert"
