+++ trace/codes/http.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package codes // import "go.opentelemetry.io/otel/codes"
+package codes // import "go.opentelemetry.io/otel/trace/codes"
 
 import (
 	"fmt"
