+++ trace/trace.go
@@ -23,8 +23,8 @@ import (
 	"strings"
 	"time"
 
-	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/label"
+	"go.opentelemetry.io/otel/trace/codes"
 )
 
 const (
